Extract useRide from RideCard
